backend/requirements.txt
asgiref==3.8.1
Bext==0.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.5
Django==5.1.1
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.6
numpy==1.26.4
packaging==24.1
pandas==2.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rpa==1.50.0
six==1.16.0
sqlparse==0.5.1
tagui==1.50.0
tzdata==2024.1
urllib3==2.2.1
